+++ presentation/src/main/java/com/sopt/presentation/appointmentCreate/appointmentCreateInfo/AppointmentCreateInfoRoute.kt
@@ -13,6 +13,9 @@ import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.BasicTextField
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -23,9 +26,13 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.TextFieldValue
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -35,7 +42,6 @@ import com.sopt.core.designsystem.component.progressbar.NoostakProgressBar
 import com.sopt.core.designsystem.component.text.NoostakHeaderText
 import com.sopt.core.designsystem.component.text.NoostakSubHeaderText
 import com.sopt.core.designsystem.component.textfield.NoostakTextField
-import com.sopt.core.designsystem.component.textfield.TimeTextField
 import com.sopt.core.designsystem.component.topappbar.NoostakTopAppBar
 import com.sopt.core.designsystem.theme.NoostakAndroidTheme
 import com.sopt.core.designsystem.theme.NoostakTheme
@@ -164,11 +170,56 @@ fun AppointmentCreateInfoScreen(
                     .fillMaxWidth()
                     .padding(top = 10.dp)
             ) {
-                TimeTextField(
-                    onValueChange = { newDuration ->
-                        appointmentDuration = newDuration
-                    }
-                )
+                var text by remember { mutableStateOf(TextFieldValue("")) }
+                var isFocused by remember { mutableStateOf(false) }
+                val isTimeExceedLimit = (text.text.toIntOrNull() ?: 0) > 10
+
+                Column(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(vertical = 15.dp)
+                ) {
+                    BasicTextField(
+                        value = text,
+                        onValueChange = { newValue ->
+                            if (newValue.text.length <= 2) {
+                                text = newValue
+                                appointmentDuration = newValue.text
+                            }
+                        },
+                        textStyle = NoostakTheme.typography.b1SemiBold.copy(textAlign = TextAlign.Right),
+                        singleLine = true,
+                        keyboardOptions = KeyboardOptions.Default.copy(
+                            keyboardType = KeyboardType.Number
+                        ),
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .padding(end = 44.dp)
+                            .onFocusChanged { focusState ->
+                                isFocused = focusState.isFocused
+                            }
+                    )
+                    HorizontalDivider(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .padding(top = 15.dp),
+                        thickness = 2.dp,
+                        color = if (isTimeExceedLimit) {
+                            NoostakTheme.colors.red02
+                        } else if (isFocused) {
+                            NoostakTheme.colors.blue600
+                        } else {
+                            NoostakTheme.colors.gray500
+                        }
+                    )
+                    Text(
+                        text = stringResource(R.string.text_calendar_appointment_time_text),
+                        style = NoostakTheme.typography.c3SemiBold,
+                        color = if (isTimeExceedLimit) NoostakTheme.colors.red02 else NoostakTheme.colors.gray500,
+                        modifier = Modifier.padding(top = 10.dp)
+                    )
+                }
+
                 Text(
                     text = stringResource(R.string.text_calendar_appointment_duration),
                     style = NoostakTheme.typography.b1SemiBold,
